Start LinkedList implementation in Java
This is a fairly straightforward implementation, except that it is a doubly-
linked list. This essentially means that there is both a "next" and "prev"
reference in each node, which allows the list to be traversed in either
direction.

It also allows for removal of indices. The current supported (but untested)
features include:
	- insertion
	- removal
	- get (by index or item)
	- contains