Fix queue resizing method to properly copy memory locations
Instead of using realloc, which does not work properly when shrinking the array,
we use malloc and memcpy. In addition, this simplifies the push/pop pointers -
the data is copied to the beginning of the new array, and thus the pop pointer
is set to index 0 and the push pointer is at the last data element.

In addition, a test task for the Makefile was added to test the implementation
when the code is built.